+++ lib/Graphics/Font.vala
@@ -4,8 +4,8 @@ using Virgil.FileSystem;
 using SDL;
 
 namespace Virgil.Graphics {
+
     public class Font {
-        private uint8[] _data;
         private SDLTTF.Font _font;
 
         public int size;
@@ -16,6 +16,7 @@ namespace Virgil.Graphics {
             }
 
             set {
+                _font.font_style = value.to_sdl ();
                 style = value;
             }
         }
@@ -26,6 +27,7 @@ namespace Virgil.Graphics {
             }
 
             set {
+                _font.hinting = value.to_sdl ();
                 hinting = value;
             }
         }
@@ -36,10 +38,43 @@ namespace Virgil.Graphics {
             }
 
             set {
+                _font.outline = value;
                 outline_size = value;
             }
         }
 
-        public Font () { }
+        public Font (string file, int size = 16) {
+            _font = new SDLTTF.Font (file, size);
+
+            if (_font == null) {
+                error (@"Font creation with file ??$file?? failed!");
+            }
+        }
+
+        public Font.from_rwops (RWops rwops, int size = 16) {
+            _font = new SDLTTF.Font.RW (rwops, 0, size);
+
+            if (_font == null) {
+                error (@"Font creation with rwops failed!");
+            }
+        }
+
+        public Font.from_asset (Asset asset, int size = 16) {
+            uint8[] data = asset.get_data ();
+
+            message (data.length.to_string ());
+
+            SDL.RWops rwops = new SDL.RWops.from_mem (data, data.length);
+
+            _font = new SDLTTF.Font.RW (rwops, 0, size);
+
+            if (_font == null) {
+                error ("Font creation with asset ??" + asset.filename + "?? failed!");
+            }
+        }
+
+        public unowned SDLTTF.Font? get_sdl_font () {
+            return _font;
+        }
     }
 }