+++ lib/Graphics/RenderQueue.vala
@@ -9,7 +9,7 @@ namespace Virgil.Graphics {
         private List<RenderQueueItem> _item_list;
         private bool _has_reversed;
 
-        private unowned Renderer renderer;
+        private unowned Renderer _renderer;
 
         public RenderQueue () {
             _item_list = new List<RenderQueueItem> ();
@@ -17,7 +17,7 @@ namespace Virgil.Graphics {
 
             _has_reversed = false;
 
-            renderer = Game.renderer.to_sdl ();
+            _renderer = Game.renderer.to_sdl ();
         }
 
         public void begin (RenderQueueMode mode = RenderQueueMode.DEFAULT, bool free_list = true) {
@@ -48,6 +48,32 @@ namespace Virgil.Graphics {
             _filter_item (item);
         }
 
+        /*
+        FIXME: This doesn't work, dont know why
+
+        Fails at line 61 due to segmentation fault despite sdl_font not returning null
+        */
+        public void draw_string (Font font, string text, Vector2 position) {
+            unowned SDLTTF.Font sdl_font = font.get_sdl_font ();
+
+            Colour colour = new Colour (255, 255, 255);
+
+            Surface surf = sdl_font.render (text, colour.get_sdl_colour ());
+
+            Texture texture = Texture.create_from_surface(_renderer, surf);
+
+            int width, height;
+
+            texture.query (null, null, out width, out height);
+
+            print (width.to_string () + " \\ " + height.to_string () + "\n");
+
+            Rectangle source = new Rectangle (0, 0, width, height);
+            Rectangle destination = new Rectangle (16, 16, width, height);
+
+            _renderer.copy (texture, source.get_sdl_rect (), destination.get_sdl_rect ());
+        }
+
         public void end () {
             if (sort_mode == RenderQueueMode.IMMEDIATE) {
                 _has_reversed = false;
@@ -66,7 +92,7 @@ namespace Virgil.Graphics {
         }
 
         private void _render_item (RenderQueueItem item) {
-            renderer.copy (item.output_texture, item.texture_rectangle, item.output_rectangle);
+            _renderer.copy (item.output_texture, item.texture_rectangle, item.output_rectangle);
         }
 
         private void _filter_item (RenderQueueItem item) {