+++ lib/Graphics/Texture2D.vala
@@ -7,12 +7,15 @@ using SDL.Video;
 namespace Virgil.Graphics {
     public class Texture2D {
         private Texture _texture;
-        private uint8[] _data;
 
         public int width { get; private set; }
         public int height { get; private set; }
 
-        public Texture2D (int width, int height, TextureType type) {
+        public bool is_locked { get; private set; }
+
+        public TextureType texture_type { get; private set; }
+
+        public Texture2D (int width, int height, TextureType type = TextureType.STREAM) {
             _texture = Texture.create (
                 Game.renderer.to_sdl (),
                 PixelRAWFormat.RGBA8888,
@@ -20,42 +23,74 @@ namespace Virgil.Graphics {
                 width, height
             );
 
-            _data = new uint8[0];
+            texture_type = type;
+            is_locked = false;
 
             _set_size (width, height);
         }
 
         public Texture2D.from_rwops (int width, int height, SDL.RWops rwops, bool free_source = false) {
-            for (int i = 0; i < rwops.size; i++) {
-                _data += rwops.read_u8 ();
-            }
 
             _texture = SDLImage.load_texture_rw (Game.renderer.to_sdl (), rwops, free_source);
-
-            _set_size (width, height);
-        }
-
-        public Texture2D.from_data (int width, int height, uint8[] data) {
-            SDL.RWops rwops = new SDL.RWops.from_mem (_data, _data.length);
-
-            _texture = SDLImage.load_texture_rw (Game.renderer.to_sdl (), rwops, true);
+            is_locked = false;
 
             _set_size (width, height);
         }
 
         public Texture2D.from_asset (int width, int height, Asset asset) {
-            _data = asset.get_data ();
+            uint8[] data = asset.get_data ();
 
-            SDL.RWops rwops = new SDL.RWops.from_mem (_data, _data.length);
+            SDL.RWops rwops = new SDL.RWops.from_mem (data, data.length);
 
             _texture = SDLImage.load_texture_rw (Game.renderer.to_sdl (), rwops, true);
+            is_locked = false;
 
             _set_size (width, height);
         }
 
-        private void _set_size (int width, int height) {
-            this.width = width;
-            this.height = height;
+        [Version (experimental_until = "0.0.5")]
+        public void set_pixels (Rectangle rectangle, void* pixels, int? pitch = null) {
+            if (texture_type == TextureType.STATIC) {
+                int width;
+
+                if (pitch == null) {
+                    width = rectangle.width * 4;
+                } else {
+                    width = pitch;
+                }
+
+                _texture.update (rectangle.get_sdl_rect (), pixels, pitch);
+            } else {
+                warning ("set_pixels () can only be used on static textures");
+            }
+        }
+
+        [Version (experimental_until = "0.0.5")]
+        public void @lock (Rectangle bounds, ref void* pixels, ref int pitch) {
+            if (is_locked) {
+                warning ("Texture is already locked");
+                return;
+            }
+
+            _texture.do_lock (
+                bounds.get_sdl_rect (),
+                out pixels,
+                out pitch
+            );
+
+            is_locked = true;
+        }
+
+        [Version (experimental_until = "0.0.5")]
+        public void @unlock () {
+            if (!is_locked) {
+                warning ("Texture is already unlocked");
+                return;
+            }
+
+            _texture.unlock ();
+
+            is_locked = false;
         }
 
         public Rectangle get_bounds () {
@@ -64,6 +99,11 @@ namespace Virgil.Graphics {
 
         public unowned Texture? get_sdl_texture () {
             return _texture;
+        }
+
+        private void _set_size (int width, int height) {
+            this.width = width;
+            this.height = height;
         }
     }
 }